updates to redstone
repeater now gives off a tiny amount of light when off, redstone ore is no longer always glowy, it has to be interacted with (like vanilla game) but is still a tiny bit emissive when off to keep it consistent with other redstone stuff, mainly redstone wires.
sculk sensor side was also updated but I'll have to work on it, it must be more consistent with other sculk blocks.
redstone block is dimmer to match recent redstone brightness changes.