Update 1.21.70
In this bulk commit, foundation for a fog overhaul has been laid, texture sets have been updated to a new format version, early support for pale garden was added and appearance of various blocks have been revamped or adjusted.

For a complete, detailed explanation of all changes in this update, refer to the changelogs section at Vanilla RTX's MCPEDL page.