Remove Blender Normals
The Blender normals haven't been updated and may just cause confusion since this wasn't directly related to Minecraft, it's best to remove it for now.

A better replacement will be coming soon, and besides, rotating normals for rendering projects should be straightforward for anyone who needs it, so in the end, this had to go.