Update 1.21.184
Minor revision for the copper age game drop support, updated textures based on the latest release, some adjustments, plus emissivity workarounds for the items that are missing a point light in the latest Previews.

As always, find the detailed changelogs of this release package since the previous on CurseForge or MCPEDL page of Vanilla RTX / Vanilla RTX Normals.